Implement auto-configuration on Linux
Unlike Windows, this implementation tries to find corresponding
characters for all available key codes, not just the common ones. The
algorithm wasn't extensively tested, but seems to be ok.

 #46

+++ KeyboardSwitch.Linux/X11/XkbDesc.cs
@@ -7,10 +7,13 @@ namespace KeyboardSwitch.Linux.X11
     internal struct XkbDesc
     {
         public IntPtr Dpy;
+
         public ushort Flags;
         public ushort DeviceSpec;
-        public ushort MinKeyCode;
-        public ushort MaxKeyCode;
+
+        public byte MinKeyCode;
+        public byte MaxKeyCode;
+
         public IntPtr Ctrls;
         public IntPtr Server;
         public IntPtr Map;
